+++ test/test_helper.rb
@@ -11,9 +11,9 @@ require 'minitest/mock'
 require 'capybara/rails'
 require 'capybara/minitest'
 require 'webmock/minitest'
-require 'selenium/webdriver'
 require 'support/rails5_assetions' # Remove once upgraded to Rails 5
-require 'database_cleaner'
+
+require 'application_system_test_case'
 
 Setting.address_processing = false
 Setting.registry_country_code = 'US'
@@ -39,41 +39,5 @@ class ActionDispatch::IntegrationTest
     WebMock.reset!
     Capybara.reset_sessions!
     Capybara.use_default_driver
-  end
-end
-
-class JavascriptIntegrationTest < ActionDispatch::IntegrationTest
-  self.use_transactional_fixtures = false
-  DatabaseCleaner.strategy = :truncation
-
-  Capybara.register_driver(:chrome) do |app|
-    options = ::Selenium::WebDriver::Chrome::Options.new
-
-    options.add_argument("--headless")
-    options.add_argument("--no-sandbox")
-    options.add_argument("--disable-dev-shm-usage")
-    options.add_argument("--window-size=1400,1400")
-
-    Capybara::Selenium::Driver.new(Rails.application, browser: :chrome, options: options)
-  end
-
-  Capybara.register_server(:silent_puma) do |app, port, _host|
-    require "rack/handler/puma"
-    Rack::Handler::Puma.run(app, Port: port, Threads: "0:2", Silent: true)
-  end
-
-  def setup
-    DatabaseCleaner.start
-
-    super
-
-    Capybara.current_driver = :chrome
-    Capybara.server = :silent_puma
-  end
-
-  def teardown
-    super
-
-    DatabaseCleaner.clean
   end
 end
